We have a set of models!
These are a little inconsistent with the the Moodle coding standards, so
we'll need to revisit this later and tidy it up. The issue with doing
this at the present time is that we haven't claimed ownership over the
tables and don't want to break the existing report_cpd module.