PHP 5.3 compatibility issue in base_model
We were making use of the ability to use an array containing a class or
(object) variable name at position 0 and a method name at position 1 as
a callable.

This behaviour was only introduced on 5.4, which is only required by
Moodle 2.7. We should avoid using it to maximise compatibility with
existing installs.

Improves #13